Move the DiagnosticsWithFix trait on the ide level
