+++ crates/ra_hir_def/src/diagnostics.rs
@@ -2,7 +2,7 @@
 
 use std::any::Any;
 
-use hir_expand::diagnostics::{Diagnostic, DiagnosticWithFix};
+use hir_expand::diagnostics::Diagnostic;
 use ra_syntax::{ast, AstPtr, SyntaxNodePtr};
 
 use hir_expand::{HirFileId, InFile};
@@ -23,13 +23,5 @@ impl Diagnostic for UnresolvedModule {
     }
     fn as_any(&self) -> &(dyn Any + Send + 'static) {
         self
-    }
-}
-
-impl DiagnosticWithFix for UnresolvedModule {
-    type AST = ast::Module;
-    fn fix_source(&self, db: &dyn hir_expand::db::AstDatabase) -> Option<Self::AST> {
-        let root = db.parse_or_expand(self.file)?;
-        Some(self.decl.to_node(&root))
     }
 }
